requirements.txt
tk
mysql-connector-python==8.0.33
